move starknet interaction to file
